refactor taxonomy
