+++ tamr_unify_client/project/categorization.py
@@ -1,6 +1,6 @@
 from tamr_unify_client.models.machine_learning_model import MachineLearningModel
-from tamr_unify_client.models.taxonomy.resource import Taxonomy
 from tamr_unify_client.project.resource import Project
+from tamr_unify_client.taxonomy.resource import Taxonomy
 
 
 class CategorizationProject(Project):
@@ -17,26 +17,26 @@ class CategorizationProject(Project):
         return MachineLearningModel(self.client, None, alias)
 
     def create_taxonomy(self, creation_spec):
-        """Creates a Taxonomy for this Categorization project.
+        """Creates a :class:`~tamr_unify_client.taxonomy.resource.Taxonomy` for this project.
 
         A taxonomy cannot already be associated with this project.
 
         :param creation_spec: The creation specification for the taxonomy, which can include name.
-        :type: dict
+        :type creation_spec: dict
         :returns: The new Taxonomy
-        :rtype: :class:`~tamr_unify_client.models.taxonomy.resource.Taxonomy`
+        :rtype: :class:`~tamr_unify_client.taxonomy.resource.Taxonomy`
         """
         alias = self.api_path + "/taxonomy"
         resource_json = self.client.post(alias, json=creation_spec).successful().json()
         return Taxonomy.from_json(self.client, resource_json, alias)
 
     def taxonomy(self):
-        """Retrieves the Taxonomy associated with Categorization project.
-
-        If a taxonomy is not already associated with this project, call create_taxonomy() first.
+        """Retrieves the :class:`~tamr_unify_client.taxonomy.resource.Taxonomy` associated with this project.
+        If a taxonomy is not already associated with this project,
+        call :func:`~tamr_unify_client.project.categorization.CategorizationProject.create_taxonomy` first.
 
         :returns: The project's Taxonomy
-        :rtype: :class:`~tamr_unify_client.models.taxonomy.resource.Taxonomy`
+        :rtype: :class:`~tamr_unify_client.taxonomy.resource.Taxonomy`
         """
         alias = self.api_path + "/taxonomy"
         resource_json = self.client.get(alias).successful().json()
