+++ tests/unit/test_taxonomy.py
@@ -8,7 +8,7 @@ from tamr_unify_client import Client
 from tamr_unify_client.auth import UsernamePasswordAuth
 from tamr_unify_client.category.collection import CategoryCollection
 from tamr_unify_client.category.resource import Category
-from tamr_unify_client.models.taxonomy.resource import Taxonomy
+from tamr_unify_client.taxonomy.resource import Taxonomy
 
 
 class TestTaxonomy(TestCase):
